Reduce precision (double to float)

+++ CUDA/fit_gpu_kernels.h
@@ -7,9 +7,9 @@
 /*  CCOPYRIGHT  */
 
 extern "C" __global__ void fit_PVM_single_kernel(	//INPUT
-							const double* 		data, 
-							const double* 		bvecs, 
-							const double* 		bvals, 
+							const float* 		data, 
+							const float* 		bvecs, 
+							const float* 		bvals, 
 							const int 		nvox, 
 							const int		ndirections,
 							const int 		nfib,
@@ -17,12 +17,12 @@ extern "C" __global__ void fit_PVM_single_kernel(	//INPUT
 							const bool 		m_include_f0, 
 							const bool		gradnonlin,
 							//INPUT-OUTPUT
-							double* 		params);
+							float* 			params);
 
 extern "C" __global__ void fit_PVM_single_c_kernel(	//INPUT
-							const double* 		data, 
-							const double* 		bvecs, 
-							const double* 		bvals, 
+							const float* 		data, 
+							const float* 		bvecs, 
+							const float* 		bvals, 
 							const int 		nvox, 
 							const int		ndirections,
 							const int 		nfib, 
@@ -32,13 +32,13 @@ extern "C" __global__ void fit_PVM_single_c_kernel(	//INPUT
 							const bool	 	m_return_fanning,
 							const bool		gradnonlin,
 							//INPUT-OUTPUT
-							double* 		params);
+							float* 			params);
 
 extern "C" __global__ void fit_PVM_multi_kernel(	//INPUT
-							const double* 		data, 
-							const double* 		params_PVM_simple_c,
-							const double* 		bvecs, 
-							const double* 		bvals, 
+							const float* 		data, 
+							const float* 		params_PVM_simple_c,
+							const float* 		bvecs, 
+							const float* 		bvals, 
 							const int 		nvox, 
 							const int		ndirections,
 							const int 		nfib, 	
@@ -46,13 +46,13 @@ extern "C" __global__ void fit_PVM_multi_kernel(	//INPUT
 							const bool 		m_include_f0,
 							const bool		gradnonlin,
 							//OUTPUT
-							double* 		params);
+							float* 			params);
 
 extern "C" __global__ void get_residuals_PVM_single_kernel(	//INPUT
-								const double* 		data, 
-								const double* 		params,
-								const double* 		bvecs, 
-								const double* 		bvals, 
+								const float* 		data, 
+								const float* 		params,
+								const float* 		bvecs, 
+								const float* 		bvals, 
 								const int 		nvox, 
 								const int		ndirections,
 								const int 		nfib, 
@@ -61,13 +61,13 @@ extern "C" __global__ void get_residuals_PVM_single_kernel(	//INPUT
 								const bool		gradnonlin,
 								const bool* 		includes_f0,								
 								//OUTPUT
-								double*			residuals);
+								float*			residuals);
 
 extern "C" __global__ void get_residuals_PVM_single_c_kernel(	//INPUT
-								const double* 		data, 
-								const double* 		params,
-								const double* 		bvecs, 
-								const double* 		bvals, 
+								const float* 		data, 
+								const float* 		params,
+								const float* 		bvecs, 
+								const float* 		bvals, 
 								const int 		nvox, 
 								const int		ndirections,
 								const int 		nfib, 
@@ -76,14 +76,14 @@ extern "C" __global__ void get_residuals_PVM_single_c_kernel(	//INPUT
 								const bool		gradnonlin,
 								const bool* 		includes_f0,								
 								//OUTPUT
-								double*			residuals);
+								float*			residuals);
 
 
 extern "C" __global__ void get_residuals_PVM_multi_kernel(	//INPUT
-								const double* 		data, 
-								const double* 		params,
-								const double* 		bvecs, 
-								const double* 		bvals, 
+								const float* 		data, 
+								const float* 		params,
+								const float* 		bvecs, 
+								const float* 		bvals, 
 								const int 		nvox, 
 								const int		ndirections,
 								const int 		nfib, 
@@ -92,5 +92,5 @@ extern "C" __global__ void get_residuals_PVM_multi_kernel(	//INPUT
 								const bool		gradnonlin,
 								const bool* 		includes_f0,								
 								//OUTPUT
-								double*			residuals);
+								float*			residuals);
